Feature/buttons (#4)
* Add Solve and Reset button elements

A Solve button and a Reset button were added to the game. Currently,
clicking either button has no functionality. The buttons change to
a darker shade when the mouse hovers over them.

* Implement Solve and Reset button functionality

The Reset button now resets the board to the starting postion for the
game. The Solve button now solves the game, but still does not
visually show the process in which the game is solved.

* Add algorithm visulaizer when Solve is pressed

The algorithm that is being used to solve the Sudoku game is now
being visualized by the GUI when the Solve button is pressed. The
cell that is currently being tested is outlined in red while all
previously tested cells are outlined in green.